+++ packages/wallet-ui/src/services/useStarkNetSnap.ts
@@ -712,15 +712,13 @@ export const useStarkNetSnap = () => {
         },
       });
       return {
-        balanceLatest: BigNumber.from(response.balanceLatest),
-        balancePending: BigNumber.from(response.balancePending),
+        balance: BigNumber.from(response.balancePending),
       };
     } catch (err) {
       //eslint-disable-next-line no-console
       console.error(err);
       return {
-        balanceLatest: BigNumber.from('0x0'),
-        balancePending: BigNumber.from('0x0'),
+        balance: BigNumber.from('0x0'),
       };
     }
   };